Import into version control
This project was developed without the intention of using Git so this is where it starts. This version of the program has the following features:
- Loading campaign data
- Importing scenes, encounters and monsters
- Navigating through scenes and encounters
- Rolling sets of dice
- Searching through scenes and encounters
However, it is lacking in some features which need to be added:
- Adding puzzles
- Some better transitions between sections (using the electron demo as a template)

There are some more things that need to be added on top of this but they will be developed in the future. Some key points to work on would be the documentation, the arrangement of the code and just some general cleanup.