Adding documentation and dividing up the structure a little bit
I've added documentation to most of the rendering thread files. Additionally, I've removed the previous renderer file and have split it into two files which are now included by the index.html file. One only handles the responses to bus events and the other handles responses to changes in the DOM. Summary changes:

~index.html: Removing require for renderer.js and adding for bus-response.js and document-response.js.
-renderer.js: Removed as it has been divided and replaced by bus-response.js and document-response.js
+bus-response.js: A renderer thread file to handle responses to bus events that mostly require interactions with the DOM (with a few exceptions such as the error dialog)
+document-response.js: a renderer thread file that deals with events occuring in the DOM.