Created the page injector
This is quite an overhall of how we are doing things. Electron wants a single page program to be written as that's what the documentation says that it expects, however, to me, this doesn't seem very intuative. I like having the different pages split up into different files. As a result I've written the page injector. This allows pages to be written as separate html files and stored in one place which will then be compiled and injected into the main page as it loads which forms a single page application despite being stored in a single place. 
There may be a few issues with how this works in the future bus as far as I can see right now, this seems to work quite well and makes things a little easier to follow for me.
I've added the ability to include css and scripts from the injected files as well which should allow for some easier expansion. 

As far as I have tested, this works and provides no real impact to speed or reactivity.