Quick visual change to the loader
I began to hate the green loader screen so I've updated it to use a plain white with black text and show the program icon. Additionally the button has been redesigned just to make it actually look nice. The old CSS is still there just in case I need it later on but it will be removed later on.