Changing the actual layout of the explorer
This is a style change for the explorer page. There is a lot of change here so I'll try and break down the non-obvious stuff.
In the index page, I've got rid of the footer (commented out for now) as I was beginning to dislike the look. I've found that the way I was dynamically adding scripts and links didn't work as jquery would not load them due to them being a sync task on the main thread so I've changed it. Now all CSS must be included in the index.html page. There will be changes coming to the CSS later on so that they can only apply to the sections they are designed to by putting the section selector before every CSS selector. This should help with any problems of overlap.

The changes to the explorer page itself form the rest of the file changes, the explorer.js is designed to handle the expanders right now as I did them separately from the material design JS.