Custom menu
I'm adding a custom menu to replace the default one from electron. This gives basic navigation features, and the ability to open a new one. I've left in the debug options too. There are entires for the new and save options for campaigns but they are not implemented right now (they are currently being worked on elsewhere).