Revising the data!
I've redesigned how that we do the data now. This is like the second time and I haven't updated any of the other implmentations of it. This will mess some things up but it doesn't have that much damage right now. I've saved the data that we had before as a backup which is why we now need the backup files in the gitignore.