Quick and dirty notification system.
This is just designed as an easy way to get short notifications to the user. Errors and stuff like that should still be done using alerts as they are more official. These are just for quick notes that are dismissed quickly such as a notification that something has been saved. 

This is very buggy as if you send a notification before the previous stuff goes to hell. (well you'll see both data but it'll just move across the screen more). I'll try and sort this at another time and maybe add a full queueing system or make it spawn new ones. I don't know yet as it will need a bit of thinking about to get sorted. I'll raise an issue in a second for it.

There's also some accidental whitespaces changes in this commit.