Fixing some failures in the scene loader
This isn't too big of a change and isn't actually done yet as we need to fix the rest of the issues with it but that will take a tiny bit more work (see issue #8). Right now, this change just supresses the errors and actually makes scene loading begin to work as it turns out whenever we tried to load a scene it would look for an encounter. I only realised this now as I changed the default from loading an encounter to a scene (huzzah for crappy programming).