Beginning the saving system.
This is the start of work on issue #6. It does the basics and allows changes from the editor to be saved to a local copy of the campaign while they are still looking at the editor page. When they hit the return button it will auto save it, reload the campaign and attempt to save it to file. 

There is a lot wrong with this right now as it does not prompt whether to save when they exit, there is no explicit save button, we have no error handling for saving past a quick notification (we should be able to ask for another save location or at least have a retry). 

The changes to the editor.js file are massive because I removed the functions from inside the fetch-campaign listener. I've also added the realSave function which actually performs the saving.