Adding save+discard buttons and change detection.
In the editor, there are now explicit save and discard buttons that the user can use. If they try and go back to the explorer without saving their changes it will give them a prompt either to save or discard.

Additionally, there is now a callback for the save function which will provide errors and stuff. We need to do a bit more with that now to actually handle the erros but that will be coming next.