Finishing up notifications for now
Notifications have been almost fully implemented now. We have not fixed the .html() vulnerability but that isn't too much of a concern as it is only internal. Icons and colors can now be specified through the options. It is worth noting that if we get the icon name wrong the notification will render very incorrectly but there's not too much to do about that right now as it is unlikely to happen. This should mark us fixing #7.