Fixing up scene loading a bit
This now makes NPCs load and display properly which was one of the main problems that we were facing. Next up is navigation but that is part of a different issue.

~ explorer.html: Adding an id to the Monsters header because it should read NPCs when displaying a scene. We are using this in bus-response.js to toggle it, hence the name.

+ npc.html: No prizes for guessing what this is. This is just the expander for NPCs with some prefilled data and ids.

~ bus-response.js: Added a function to generate the NPCs from npc.html and linked it in with the scene loading. Added lines in the scene and the encounter loader to switch that header that was mentioned earlier.

These changes still need commenting but that is part of #9. This may not completely resolve everything in #8 but we are going to mark it closed for now as it seems to work. Navigation is still broken but that is addressed by #5 instead. Therefore: fixes #8.