Fixing up entity navigation
So in our old data format we had nav specified with a location property and an optional icon. Since we've removed the option for icon and switched to a key value system everything was breaking a little bit.

To fix this we remove checks for location and change all references to a navigations location to just the value instead. This cleaned everything up and made it all work.
However, during testing I realised that something was a bit wrong when I switched a few times using the nav buttons and found about 20 puzzles. Turns out we overlooked removing the previous handlers from the nav buttons but that's fixed now so everything happens as it should.
I've also copied down the navigation code from encounter to make the top right next and previous buttons work when viewing a scene. 
Also I kept seeing multiple puzzles when I shouldnt and it turns out that we used the wrong id so scenes weren't actually clearing the puzzle container. That's all fixed. Damn plurals.