Adding expanders for encounters and scenes
As the title implies, these are just basic expanders for encounters and scenes. These are designed to be shown in the search pane only and will only list the ids of the puzzles and stuff linked with them for now. We may expand it to include their names too but we will do that later on.