Searching!
Searching is now actually possible. It's not the perfect system so I will be tweaking it a little bit in the future. I'll raise an issue about it as a central place to note problems and stuff like that.

It's a bit of a mess right now but bear with me. Generated elements that are inserted are injected with either the prefix 'ex-srch-' or 'ex-srcha-' depending on whether they are in the all section. This is then used to make them appear or disappear as the user searches. Additionally, they have a letter following them to prevent duplicate ids.

There are a load of formatting changes that are mixed in with this which will make this commit look more complicated than it actually is.

I've added a word indexing system so as each expander is generated, all words used in it will be indexed with its id. If the word has already been indexed then the id will be added to the existing id(s). This means that you can search a word and return every one that mentions that word. It's not perfect but it works which is pretty cool, first time I've done something like it.

I went through and added the suffixes to all the index calls so that they'd respond to the right elements in the future.

The search function is a bit of a mess, I have to admit. It searches for all occurances of each word in the array of strings given to it. Finds every one in which at least one of the words appears then orders them according to the number of matches. This doesn't actually do anything right now but its designed so later on we can rearrange the list to show the most accurate results at the top but that will come later on as it requires reordering dom elements which I don't want to think about to be honest.

The registering of the search fields basically take the input and search with it split at spaces. This means that any incomplete words will not return any results. Not, perfect but there we go. It then just shows the ones that match and hides the ones that don't.