Ordering of search results
This is a bit messy but I'm basically abusing how append works. We have the order that they should be in based on the search function so now we hide every element and then detach and append each of the matches in order. As they are appended at the bottom of their parent, this means that the list should be in the right order. From my quick testing they seem to be in the same order as the result from search so this should work. It's not too major if it doesn't work anyway. It's just a nice feature. This provides a fix for one of the boxes under #11 as far as I know.